facebook_poster/facebook_automation/requirements.txt
crewai==0.28.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.1
facebook-sdk==3.1.0
schedule==1.2.0